pyproject.toml
[project]
name = "polydiff"
version = "0.1.0"
description = ""
authors = [
    {name = "Jerenemy",email = "jzay@wesleyan.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.4.2,<3.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "torch (>=2.2.0,<3.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0.0,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
